refactor(cli): rename config to storeConfig
