+++ packages/cli/src/utils/tablegen.ts
@@ -6,7 +6,7 @@ import {
   RenderSchemaType,
 } from "../contracts/renderSchema.js";
 import { SchemaType, SchemaTypeArrayToElement, SchemaTypeId, getStaticByteLength } from "@latticexyz/schema-type";
-import { StoreConfig } from "./config.js";
+import { StoreConfig } from "./storeConfig.js";
 
 export function renderTables(config: StoreConfig) {
   const storeImportPath = config.storeImportPath;
